refactor: distinguished auth and car in member server
