+++ src/main/java/com/fullcar/core/config/jwt/JwtTokenProvider.java
@@ -1,6 +1,6 @@
 package com.fullcar.core.config.jwt;
 
-import com.fullcar.member.domain.Member;
+import com.fullcar.member.domain.member.Member;
 import io.jsonwebtoken.*;
 import io.jsonwebtoken.io.Decoders;
 import io.jsonwebtoken.security.Keys;
@@ -8,12 +8,10 @@ import jakarta.servlet.http.HttpServletRequest;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Value;
-import org.springframework.security.core.Authentication;
 import org.springframework.stereotype.Component;;
 
 import java.security.Key;
 import java.util.Date;
-import java.util.Objects;
 
 @Slf4j
 @Component
