--- src/main/java/com/fullcar/member/application/AuthService.java
@@ -1,8 +1,0 @@
-package com.fullcar.member.application;
-
-import com.fullcar.member.presentation.dto.request.AuthRequestDto;
-import com.fullcar.member.presentation.dto.response.SocialInfoResponseDto;
-
-public interface AuthService {
-    SocialInfoResponseDto getMemberInfo(AuthRequestDto authRequestDto);
-}
