+++ src/main/java/com/fullcar/member/application/auth/KakaoAuthService.java
@@ -1,4 +1,4 @@
-package com.fullcar.member.application;
+package com.fullcar.member.application.auth;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.JsonNode;
@@ -6,10 +6,14 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fullcar.core.config.jwt.JwtTokenProvider;
 import com.fullcar.core.exception.NotFoundException;
 import com.fullcar.core.response.ErrorCode;
-import com.fullcar.member.domain.*;
-import com.fullcar.member.presentation.dto.KakaoInfoDto;
-import com.fullcar.member.presentation.dto.request.AuthRequestDto;
-import com.fullcar.member.presentation.dto.response.SocialInfoResponseDto;
+import com.fullcar.member.domain.auth.SocialId;
+import com.fullcar.member.domain.auth.service.SocialIdService;
+import com.fullcar.member.domain.member.Member;
+import com.fullcar.member.domain.member.service.MemberIdService;
+import com.fullcar.member.domain.member.MemberRepository;
+import com.fullcar.member.presentation.auth.dto.KakaoInfoDto;
+import com.fullcar.member.presentation.auth.dto.request.AuthRequestDto;
+import com.fullcar.member.presentation.auth.dto.response.SocialInfoResponseDto;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpEntity;
