+++ src/main/java/com/fullcar/member/domain/member/MemberId.java
@@ -1,4 +1,4 @@
-package com.fullcar.member.domain;
+package com.fullcar.member.domain.member;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Embeddable;
