+++ src/main/java/com/fullcar/member/presentation/auth/MemberController.java
@@ -1,4 +1,4 @@
-package com.fullcar.member.presentation;
+package com.fullcar.member.presentation.auth;
 
 import org.springframework.web.bind.annotation.RestController;
