+++ src/main/java/com/fullcar/member/presentation/auth/dto/KakaoInfoDto.java
@@ -1,4 +1,4 @@
-package com.fullcar.member.presentation.dto;
+package com.fullcar.member.presentation.auth.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
