+++ src/main/java/com/fullcar/member/presentation/auth/dto/response/AuthTokenResponseDto.java
@@ -1,4 +1,4 @@
-package com.fullcar.member.presentation.dto.response;
+package com.fullcar.member.presentation.auth.dto.response;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.Builder;
